### 2022-08-07 #4542 @majensen
Hi  @ethanwhite - would you be available and willing to review [this work](#4542) for [JOSS](https://joss.theoj.org)? I would appreciate your Python data munging expertise. Thanks! --@majensen
