### 2022-08-15 #4542 @majensen
@svburke many thanks for agreeing to review [this work](#4542) for [JOSS](https://joss.theoj.org)! I will add you as a reviewer now.
