### 2022-08-15 #4678 @editorialbot
[REVIEW]: fars_cleaner:  A Python package for downloading and pre-processing vehicle fatality data in the US
**Submitting author:** @mzabrams (<a href="http://orcid.org/0000-0001-6818-5214">Mitchell Abrams</a>)
**Repository:** https://github.com/mzabrams/fars-cleaner
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.5
