### 2022-08-28 mzabrams/fars-cleaner#1 @svburke
pandas.errors.ParserError: Error tokenizing data. C error: out of memory
While running the example command given here:

```
from fars_cleaner import FARSFetcher, load_pipeline

fetcher = FARSFetcher(project_dir=SOME_PATH)
vehicles, accidents, people = load_pipeline(fetcher=fetcher,
                                            first_run=True,
                                            target_folder=SOME_PATH)
```

I am getting an error on one of my machines (Windows 10 16Gb RAM):

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\fars_cleaner\data_loader.py", line 128, in load_pipeline
    vehicle, person, accident = load_basic(year,
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\fars_cleaner\data_loader.py", line 523, in load_basic
    veh_df = pd.read_csv(vehicle_file,
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\util\_decorators.py", line 311, in wrapper
    return func(*args, **kwargs)
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 680, in read_csv
    return _read(filepath_or_buffer, kwds)
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 581, in _read
    return parser.read(nrows)
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 1255, in read
    index, columns, col_dict = self._engine.read(nrows)
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\c_parser_wrapper.py", line 230, in read
    data = self._reader.read(nrows)
  File "pandas\_libs\parsers.pyx", line 787, in pandas._libs.parsers.TextReader.read
  File "pandas\_libs\parsers.pyx", line 876, in pandas._libs.parsers.TextReader._read_rows
  File "pandas\_libs\parsers.pyx", line 1960, in pandas._libs.parsers.raise_parser_error
pandas.errors.ParserError: Error tokenizing data. C error: out of memory
```

I was able to successfully run this on a different machine, (MacOS Monterey 12.5.1, 16 Gb RAM).

I think it would be worth looking into a way to read in and handle the data in a less memory intense way.
