### 2022-08-29 mzabrams/fars-cleaner#2 @svburke
Mappings file
For the last part of the usage README, you have the following code:

```
from fars_cleaner.data_loader import load_basic

vehicles, accidents, people = load_basic(year=1975, data_dir=SOME_PATH, mapper=mappings)
```

When running this code, pointing to the given "mapping.dict" file, I get the error:

```
TypeError: load_basic(), got an unexpected keyword argument 'mapper'
```

Checked the code, https://github.com/mzabrams/fars-cleaner/blob/master/fars_cleaner/data_loader.py#L469-L487, you need to have "mapping", not "mapper" for your example code.

After changing that in the code and running it again, I got this error:

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/burkesv/env/lib/python3.9/site-packages/fars_cleaner/data_loader.py", line 474, in load_basic
     acc_cols = get_renaming(mapping['Accident'], year)
TypeError: string indices must be integers
```
