### 2022-08-29 #4678 @svburke
I have left the following two issues on the main repository:

https://github.com/mzabrams/fars-cleaner/issues/1
https://github.com/mzabrams/fars-cleaner/issues/2
