### 2022-08-29 26751f2
Update example code, fix Issue #2
fix #2
