### 2022-08-29 mzabrams/fars-cleaner#2 @mzabrams
Thanks for catching this -- I've updated the README to fix that
