### 2022-08-29 mzabrams/fars-cleaner#1 @mzabrams
Interesting -- not sure why that would throwing an out of memory error on Windows only, I'll look into that
