### 2022-09-05 #4678 @mzabrams
@svburke I've addressed one of those two issues, and I am examining the other ([mzabrams/fars-cleaner#1](https://github.com/mzabrams/fars-cleaner/issues/1)) -- I'm not sure why this is the behavior on windows and not on macOS, but I am trying to assess ways of importing the full dataset without running into the out-of-memory error on a 16GB setup. Further thoughts/progress will be addressed in the issue on the project repository as I work through it.
