### 2022-09-05 mzabrams/fars-cleaner#1 @mzabrams
@svburke Would you be able to rerun this with the debug flag set to 2 in `load_pipeline()`, and post the result with how many years are loaded prior to the out of memory error?
