### 2022-09-06 mzabrams/fars-cleaner#4 @ethanwhite
Expand community guidelines to including reporting and seeking support
The JOSS guiidlines state that software should include community guidelines that provide information:

> for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

The current README directs provides a brief statement about (1), but not (2) or (3). The README could be briefly expanded to include these two additional points.

In addition it might be worth adding a short `contributing.md` file following the instructions here: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file

This is a review issue for https://github.com/openjournals/joss-reviews/issues/4678
