### 2022-09-06 mzabrams/fars-cleaner#6 @ethanwhite
Documentation missing for `load_basic` making `mapping` argument unclear
`load_basic` is presented in the README as a user facing function, but the function is undocumented. While the first three arguments can be understood by looking at the documentation for `load_pipeline` and the README, the only source of information I could find for `mapping` was by reading the code.

Related to this is that using the default value for `mapping` results in an error:

```python
from fars_cleaner.data_loader import load_basic

fetcher = FARSFetcher(project_dir=Path("fars-test/"))
fetcher.fetch_all()
vehicles, accidents, people = load_basic(year=2020, data_dir="fars-test/")
```

Result:

```
Input In [42], in <cell line: 1>()
----> 1 vehicles, accidents, people = load_basic(year=2020, data_dir=Path("fars-test/"))

File ~/miniconda3/envs/fars-cleaner-test/lib/python3.10/site-packages/fars_cleaner/data_loader.py:474, in load_basic(year, use_dask, data_dir, mapping)
    472 person_file = cur_year / "PERSON.csv"
    473 accident_file = cur_year / "ACCIDENT.csv"
--> 474 acc_cols = get_renaming(mapping['Accident'], year)
    475 per_cols = get_renaming(mapping['Person'], year)
    476 veh_cols = get_renaming(mapping['Vehicle'], year)

TypeError: 'NoneType' object is not subscriptable
```

So better documentation and either a working default or making the argument a required argument would be helpful for clarifying proper usage.

This is a review issue for https://github.com/openjournals/joss-reviews/issues/4678
