### 2022-09-07 #4678 @ethanwhite
Overall things look good. I've opened 4 issues in the fars-cleaner repo related to installation, documentation, functionality, and community guidelines, all of which are linked above.
