### 2022-09-17 mzabrams/fars-cleaner#1 @svburke
@mzabrams, here is the first output from the load_pipeline() with the debug=2 flag:

```
Fetched originals, loading to pandas....
1975
1976
1977
1978

Traceback (most recent call last):
File "<stdin>", line 1, in <module>
 File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\fars_cleaner\data_loader.py", line 128, in load_pipeline
 vehicle, person, accident = load_basic(year,
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\fars_cleaner\data_loader.py", line 531, in load_basic
per_df = pd.read_csv(person_file, encoding='cp1252',
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\util\_decorators.py", line 311, in wrapper
return func(*args, **kwargs)
  File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 680, in read_csv
 return _read(filepath_or_buffer, kwds)
File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 581, in _read
return parser.read(nrows)
 File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\readers.py", line 1255, in read
index, columns, col_dict = self._engine.read(nrows)
   File "C:\Users\svbur\AppData\Local\Programs\Python\Python38-32\lib\site-packages\pandas\io\parsers\c_parser_wrapper.py", line 230, in read
data = self._reader.read(nrows)
   File "pandas\_libs\parsers.pyx", line 787, in pandas._libs.parsers.TextReader.read
  File "pandas\_libs\parsers.pyx", line 876, in pandas._libs.parsers.TextReader._read_rows
File "pandas\_libs\parsers.pyx", line 1960, in pandas._libs.parsers.raise_parser_error
pandas.errors.ParserError: Error tokenizing data. C error: out of memory
```

I tried running it a few more times, and it would then crash on 1977. Same error output still.
