### 2022-09-26 8582829
Update docstring for load_basic, add CONTRIBUTING.md
