+++ fars_cleaner/data_loader.py
@@ -58,6 +58,7 @@ def load_pipeline(
         Flag to determine whether to parallelize using Dask. Supported, but not recommended at this time.
         The default is False.
     client : Dask Distributed Client, optional. Required if use_dask is True.
+        When provided and use_dask is False, use_dask is set to True.
     fetcher : FARSFetcher, optional
         FARSFetcher instance to facilitate file download and cache management. Overrides `load_from` value.
     debug : int, optional
@@ -81,6 +82,9 @@ def load_pipeline(
 
     if use_dask and client is None:
         raise ValueError("Must provide a client to use Dask.")
+
+    if client is not None and not use_dask:
+        use_dask = True
 
     if fetcher is None:
         if load_from is None:
@@ -128,7 +132,8 @@ def load_pipeline(
             vehicle, person, accident = load_basic(year,
                                                    use_dask=use_dask,
                                                    data_dir=data_dir,
-                                                   mapping=mappers)
+                                                   mapping=mappers,
+                                                   client=client)
             accident['YEAR'] = year
             vehicle['YEAR'] = year
             person['YEAR'] = year
@@ -136,6 +141,14 @@ def load_pipeline(
             lazy_people.append(person)
             lazy_vehicles.append(vehicle)
             lazy_accidents.append(accident)
+            if debug > 1:
+                #print('dense: acc {:0.2f} bytes, per {:0.2f} bytes, veh {:0.2f} bytes'.format(accident.memory_usage(), person.memory_usage(deep=True), vehicle.memory_usage(deep=True)))
+                print(accident.info())
+                print(vehicle.info())
+                print(person.info())
+            del accident
+            del person
+            del vehicle
 
         if use_dask:
             people = dd.concat(lazy_people)
@@ -419,6 +432,8 @@ def process_people(people, mappers):
                 'P_SF2': 'P_CF2',
                 'P_SF3': 'P_CF3',
             }))
+    per = per.encode_categorical(list(
+        {'STATE', 'SEX', 'PER_TYP', 'SEAT_POS', 'P_CF1', 'P_CF2', 'P_CF3'} & set(per.columns)))
 
     return createPerID(per, None)
 
@@ -466,7 +481,41 @@ def mapping(group, mappers):
     return group.replace(cur_mappers)
 
 
-def load_basic(year, use_dask=True, data_dir=None, mapping=None):
+def load_basic(year, use_dask=False, data_dir=None, mapping=None, client=None):
+    """Basic load pipeline to load FARS data for `year`
+
+    Assumes FARS data for a given year is already downloaded and unzipped in `data_dir`, and
+    requires a mapping dictionary to appropriately rename fields within the relevant dataframes.
+
+
+    Parameters
+    ----------
+    year : int
+        Year of data to load
+    data_dir : Path-like
+        Path where the data is stored. Must contain the folder "YYYY.unzip", where YYYY=`year`
+    use_dask : bool, optional
+        Flag to determine whether to parallelize using Dask. Supported, but not recommended at this time.
+        The default is False.
+    client : Dask Distributed Client, optional. Required if use_dask is True.
+    mapping: dict
+        mapping dictionary, as generated by fars_cleaner.builder.load_sheets()
+
+    Returns
+    -------
+    (pandas.DataFrame, pandas.DataFrame, pandas.DataFrame)
+        Returns three DataFrames: vehicles, accidents, and people.
+
+    Raises
+    ------
+    ValueError
+        If `use_dask` is True and no client provided.
+
+    """
+
+    if use_dask and client is None:
+        raise ValueError("Must provide a client to use Dask.")
+
     cur_year = data_dir / f"{year}.unzip"
     vehicle_file = cur_year / "VEHICLE.csv"
[…]
