+++ CONTRIBUTING.md
@@ -1,0 +1,36 @@
+# Contributing to fars-cleaner
+
+Have something to contribute? Found an issue? We'd love to know! This file has
+some guidelines on sharing improvements, issues, ideas, or questions with the 
+developers. In general, any contribution should start with opening an issue on
+this repository. These guidelines are heavily influenced by the Ruby 
+on Rails [contribution guidelines](https://github.com/rails/rails/blob/main/CONTRIBUTING.md).
+
+## Contributions
+This software is licensed under the BSD-3 Clause license. Any code changes you
+submit will fall under this license as well.
+
+## How to contribute
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/mzabrams/fars-cleaner/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/mzabrams/fars-cleaner/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+#### **Did you fix whitespace, format code, or make a purely cosmetic patch?**
+
+Changes that are cosmetic in nature and do not add anything substantial to the stability, functionality, or testability of this package will generally not be accepted.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Open an issue to discuss the new feature. Feel free to fork the repository and work on your improvement!
+
+#### **Do you have questions about the source code or how to use this package?**
+
+* Questions are welcome -- please contact the maintainers directly, or open an issue.
