### 2022-09-26 cbbcf52
Remove unneeded and unused methods
