+++ fars_cleaner/fars_utils.py
@@ -11,20 +11,6 @@ import pandas as pd
 import numpy as np
 import requests, json
 
-
-def get_vin_info(vin, s = requests.Session()):
-    url = 'https://vpic.nhtsa.dot.gov/api/vehicles/DecodeVinValuesExtended/'
-    fields = {'format': 'json', 'data': vin}
-    resp = s.post(url, data=fields)
-    result = resp.json().get('Results')
-    return {vin: dict(result)}
-
-
-def standard_error(R, a, b, c, d, sigma_u = 0.05):
-    return R*np.sqrt(sigma_u**2 + 1/a + 1/b + 1/c + 1/d)
-    # We take sigma_u = intrinsic uncertainty, .05
-
-
 def createVehID(df, yr):
     x = df
 
