### 2022-09-26 0c6eec2
Create test_data_loader.py
