+++ meta.yaml
@@ -1,22 +1,48 @@
+{% set name = "fars-cleaner" %}
+{% set version = "1.2.0" %}
+
 package:
-  name: fars_cleaner
-  version: "1.2.0"
+  name: {{ name|lower }}
+  version: {{ version }}
 
 source:
-  git_rev:
-  git_url: https://github.com/mzabrams/fars-cleaner
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/fars_cleaner-{{ version }}.tar.gz
+  sha256: a9421abf1ada4c801e82a5b84c6ceb7327ab32d61cb546dfa8ea11375811d4cf
+
+build:
+  noarch: python
+  script: {{ PYTHON }} -m pip install . -vv
+  number: 0
 
 requirements:
-  build:
-    - python
-    - setuptools
-
+  host:
+    - pip
+    - python >=3.7
   run:
-    - python
-
+    - dask-core
+    - numpy
+    - pandas
+    - pathlib
+    - pooch >=1.6.0
+    - pyjanitor >=0.23.1
+    - python >=3.7
+    - requests
+    - tqdm
+    - fuzzyfinder
 test:
   imports:
-    -
+    - fars_cleaner
+  commands:
+    - pip check
+  requires:
+    - pip
 
 about:
-  home:
+  home: https://pypi.org/project/fars-cleaner/
+  summary: A package for loading and preprocessing the NHTSA FARS crash database
+  license: BSD-3-Clause
+  license_file: LICENSE
+
+extra:
+  recipe-maintainers:
+    - mzabrams
