### 2022-09-26 64fc2e9
Remove debug print statement
