+++ environment.yml
@@ -7,6 +7,7 @@ dependencies:
   - numpy
   - pandas
   - dask
+  - distributed
   - tqdm
   - pooch >= 1.6.0
   - pathlib
