### 2022-09-29 mzabrams/fars-cleaner#5 @mzabrams
Implemented fuzzy-finder matching  to find appropriate files (should be futureproofed as well)
