### 2022-09-29 mzabrams/fars-cleaner#1 @mzabrams
I've made some progress on reducing memory overhead on a 16GB Windows 10 VM, but haven't been able to fully execute this program under those specifications. Work in progress, but may require a note in the README on system requirement limitations for the time being.
