### 2022-10-04 #4678 @mzabrams
@majensen I've made some progress on these -- two issues remain open, one may need to be closed with a note made to system requirements to run the package (there seems to be a Windows-based issue that I've been unable to resolve to this point, potentially based on the way Windows handles memory usage compared to macOS and Linux systems), and I'm working on fixing the packaging for download from conda/conda-forge. Hopefully by the end of the week I'll have all this sorted!
