+++ pyproject.toml
@@ -1,16 +1,45 @@
 # pyproject.toml
 
 [build-system]
-requires = ["setuptools>=61.0.0", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
 
-[project]
+[tool.poetry.dependencies]
+python = "^3.7"
+pooch = ">=1.6.0"
+pandas = "^1.0.0"
+numpy = "^1.22.0"
+pathlib = "*"
+pyjanitor = "^0.23.1"
+dask = "*"
+distributed = "*"
+requests = "*"
+tqdm = "*"
+fuzzyfinder = "*"
+
+pytest = {version = "^7.1.0", optional = true}
+hypothesis = {version = "*", optional = true}
+
+[tool.poetry]
 name = "fars_cleaner"
 version = "1.2.0"
-authors= [{ name = "Mitchell Abrams", email = "mitchell.abrams@duke.edu" }]
-license = { file = "LICENSE" }
+authors= [
+    "Mitchell Abrams <mitchell.abrams@duke.edu>",
+]
+maintainers = [
+    "Mitchell Abrams <mitchell.abrams@duke.edu>",
+]
+license = "BSD-3-Clause"
 readme = "README.md"
 description = "A package for loading and preprocessing the NHTSA FARS crash database"
+repository = "https://github.com/mzabrams/fars-cleaner"
+keywords = [
+    "FARS",
+    "crash analysis",
+    "data preprocessing",
+    "NHTSA",
+    "vehicle safety"
+]
 classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
@@ -19,33 +48,13 @@ classifiers = [
     "License :: OSI Approved :: BSD License",
     "Development Status :: 5 - Production/Stable",
 ]
-keywords = [
-    "FARS",
-    "crash analysis",
-    "data preprocessing",
-    "NHTSA",
-    "vehicle safety"
-]
-dependencies = [
-    "pooch >= 1.6.0",
-    "pandas",
-    "numpy",
-    "pathlib",
-    "pyjanitor >= 0.23.1",
-    "dask",
-    "distributed",
-    "requests",
-    "tqdm",
-    "fuzzyfinder"
-]
-requires-python = ">=3.7"
+homepage = "https://github.com/mzabrams/fars-cleaner"
+#packages = [{include = "poetry_demo"}]
 
-[tool.setuptools]
-packages = ["fars_cleaner"]
-include-package-data = true
 
-[project.optional-dependencies]
+#[tool.setuptools]
+#packages = ["fars_cleaner"]
+#include-package-data = true
+
+[tool.poetry.extras]
 dev = ["pytest", "hypothesis"]
-
-[project.urls]
-"Homepage" = "https://github.com/mzabrams/fars-cleaner"
