+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "poetry.core.masonry.api"
 python = "^3.7"
 pooch = ">=1.6.0"
 pandas = "^1.0.0"
-numpy = "^1.22.0"
+numpy = "^1.21.0"
 pathlib = "*"
 pyjanitor = "^0.23.1"
 dask = "*"
