+++ .github/workflows/ci.yml
@@ -24,4 +24,4 @@ jobs:
       - name: Install Dependencies
         run: poetry install
       - name: Test with pytest
-        run: poetry run pytest 
+        run: poetry run pytest
