+++ .github/workflows/ci.yml
@@ -9,13 +9,14 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-12, macos-11, windows-latest ]
-
+        python-version: [ '3.7', '3.8', '3.9' ,'3.10']
+      fail-fast: false
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: '3.x'
+          python-version: ${{ matrix.python-version }}
       - name: Install Poetry
         uses: snok/install-poetry@v1
         with:
