### 2022-10-06 a9a4401
Update versioning for numpy, CI install
