+++ .github/workflows/ci.yml
@@ -23,6 +23,6 @@ jobs:
           virtualenvs-create: true
           virtualenvs-in-project: true
       - name: Install Dependencies
-        run: poetry install
+        run: poetry install --extras "dev"
       - name: Test with pytest
         run: poetry run pytest
