+++ pyproject.toml
@@ -7,11 +7,13 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry.dependencies]
 python = "^3.7"
 pooch = ">=1.6.0"
-pandas = "^1.0.0"
+pandas = "^1.4.0"
 numpy = [
     {version = "^1.21.0", python = "<3.8"},
     {version = "^1.22.0", python = ">=3.8"},
 ]
+scipy = "^1.7"
+
 pathlib = "*"
 pyjanitor = "^0.23.1"
 dask = "*"
