### 2022-10-10 b69295d
Fix pandas dep versioning for 3.7
