+++ pyproject.toml
@@ -7,7 +7,10 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry.dependencies]
 python = "^3.7"
 pooch = ">=1.6.0"
-pandas = "^1.4.0"
+pandas = [
+    {version = "^1.3", python = "<3.8"},
+    {version = "^1.4", python = ">=3.8"}
+]
 numpy = [
     {version = "^1.21.0", python = "<3.8"},
     {version = "^1.22.0", python = ">=3.8"},
