### 2022-10-10 c3af34a
Fix for scipy
