+++ pyproject.toml
@@ -15,7 +15,7 @@ numpy = [
     {version = "^1.21.0", python = "<3.8"},
     {version = "^1.22.0", python = ">=3.8"},
 ]
-scipy = "^1.7"
+scipy = "^1.7.0"
 
 pathlib = "*"
 pyjanitor = "^0.23.1"
