+++ pyproject.toml
@@ -5,14 +5,12 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 pooch = ">=1.6.0"
 pandas = [
-    {version = "^1.3", python = "<3.8"},
     {version = "^1.4", python = ">=3.8"}
 ]
 numpy = [
-    {version = "^1.21.0", python = "<3.8"},
     {version = "^1.22.0", python = ">=3.8"},
 ]
 scipy = "^1.7.0"
