+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-12, macos-11, windows-latest ]
-        python-version: [ '3.7', '3.8', '3.9' ,'3.10']
+        python-version: ['3.8', '3.9' ,'3.10']
       fail-fast: false
     steps:
       - uses: actions/checkout@v3
