+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python >= 3.7
+  - python >= 3.8
   - numpy
   - pandas
   - dask
