### 2022-10-10 9eecd18
Pin version of Poetry installer to fix Windows 3.8 issue
