+++ .travis.yml
@@ -7,12 +7,14 @@ before_install:
   - pip install -U pip
   - pip install -U pytest
   - pip install coverage
+  - pip install poetry
 install:
-  - pip install -U ".[dev]"
+  #- pip install -U ".[dev]"
   #- pip install -r requirements.txt
-  - pip install coverage
+  #- pip install coverage
+  - poetry install --extras "dev"
 # command to run tests
 script:
-  - coverage run -m pytest
+  - poetry run coverage run -m pytest
 after_success:
   - bash <(curl -s https://codecov.io/bash)
