+++ .travis.yml
@@ -3,6 +3,7 @@ dist: focal
 python:
   - "3.8"
   - "3.9"
+  - "3.10"
 # command to install dependencies
 before_install:
   - pip install -U pip
