+++ pyproject.toml
@@ -22,6 +22,7 @@ distributed = "*"
 requests = "*"
 tqdm = "*"
 fuzzyfinder = "*"
+tornado = "^6.2"
 
 pytest = {version = "^7.1.0", optional = true}
 hypothesis = {version = "*", optional = true}
