+++ .github/workflows/ci.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-latest, macos-12, macos-11, windows-latest ]
+        os: [ ubuntu-latest, macos-12, macos-11, windows-latest, self-hosted ]
         python-version: ['3.8', '3.9' ,'3.10']
       fail-fast: false
     steps:
