### 2022-10-17 d7ccfde
Update CI shell to bash for windows
