+++ .github/workflows/ci.yml
@@ -11,6 +11,7 @@ jobs:
         os: [ ubuntu-latest, macos-12, macos-11, windows-latest, self-hosted ]
         python-version: ['3.8', '3.9' ,'3.10']
       fail-fast: false
+      shell: bash
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
