### 2022-10-17 50628eb
Remove self-hosted runner
