+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-latest, macos-12, macos-11, windows-latest, self-hosted ]
+        os: [ ubuntu-latest, macos-12, macos-11, windows-latest]
         python-version: ['3.8', '3.9' ,'3.10']
       fail-fast: false
     steps:
