### 2022-10-17 b0afdd7
Try adding macos and windows tests to Travis
