+++ .github/workflows/ci.yml
@@ -24,6 +24,7 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1.3.1
         with:
+          version: 1.2.1
           virtualenvs-create: true
           virtualenvs-in-project: true
       - name: Install Dependencies
