### 2022-10-17 5cbae36
Add homebrew to travis macOS
