+++ .travis.yml
@@ -14,8 +14,13 @@ jobs:
       os: osx
       osx_image: xcode11.3  # Python 3.7.4 running on macOS 10.14.4
       language: shell       # 'language: python' is an error on Travis CI macOS
+      addons:
+        homebrew:
+          packages:
+            - python@3.8
       before_install:
         - python3 --version
+        - python --version
         - python3 -m pip install --upgrade pip
         - python3 -m pip install --upgrade codecov pytest poetry
     # ------------ Windows JOBS --------------------
