+++ .github/workflows/ci.yml
@@ -12,8 +12,15 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-latest, macos-12, macos-11, windows-latest]
+        os: [ ubuntu-20.04, ubuntu-22.04, macos-12, macos-11, windows-latest]
         python-version: ['3.8', '3.9' ,'3.10']
+        include:
+          - os: windows-latest
+            continue-on-error: true
+          - os: ubuntu-20.04
+            continue-on-error: true
+          - os: ubuntu-22.04
+            continue-on-error: true
       fail-fast: false
     steps:
       - uses: actions/checkout@v3
