### 2022-10-17 a9de9e8
Pin poetry to 1.2.1 in travis
