+++ .travis.yml
@@ -9,7 +9,7 @@ before_install:
   - pip install -U pip
   - pip install -U pytest
   - pip install coverage
-  - pip install poetry
+  - pip install poetry==1.2.1
 install:
   #- pip install -U ".[dev]"
   #- pip install -r requirements.txt
