### 2022-10-17 mzabrams/fars-cleaner#3 @mzabrams
Track issue at [conda-forge/pyjanitor-feedstock#57](https://github.com/conda-forge/pyjanitor-feedstock/pull/57)
