### 2022-10-17 mzabrams/fars-cleaner#1 @mzabrams
Closing this as "won't fix" -- this seems to be OS dependent and I've been unsuccessful in fixing this issue to date. A note has been added to the README regarding this issue.
