+++ pyproject.toml
@@ -28,7 +28,7 @@ hypothesis = {version = "*", optional = true}
 
 [tool.poetry]
 name = "fars_cleaner"
-version = "1.2.0"
+version = "1.3.0"
 authors= [
     "Mitchell Abrams <mitchell.abrams@duke.edu>",
 ]
