+++ meta.yaml
@@ -1,7 +1,7 @@
 #{% set pyproject = load_file_data("pyproject.toml", ) %}
 #{% set poetry = pyproject.get('tool', {}).get('poetry') %}
 {% set name = "fars_cleaner" %}
-{% set version = "1.2.0" %}
+{% set version = "1.3.0" %}
 
 package:
   name: {{ name|lower }}
