### 2022-10-18 7102f73
Skip test cases that fail on Github runners, update meta and pyproject.yaml, bump to v1.3.1
