+++ pyproject.toml
@@ -28,7 +28,7 @@ hypothesis = {version = "*", optional = true}
 
 [tool.poetry]
 name = "fars_cleaner"
-version = "1.3.0"
+version = "1.3.1"
 authors= [
     "Mitchell Abrams <mitchell.abrams@duke.edu>",
 ]
@@ -64,3 +64,5 @@ homepage = "https://github.com/mzabrams/fars-cleaner"
 
 [tool.poetry.extras]
 dev = ["pytest", "hypothesis"]
+
+
