+++ meta.yaml
@@ -1,38 +1,30 @@
-#{% set pyproject = load_file_data("pyproject.toml", ) %}
-#{% set poetry = pyproject.get('tool', {}).get('poetry') %}
 {% set name = "fars_cleaner" %}
-{% set version = "1.3.0" %}
+{% set version = "1.3.1" %}
 
 package:
   name: {{ name|lower }}
   version: {{ version }}
 
 source:
-  path: .
-  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/fars_cleaner-{{ version }}.tar.gz
-  sha256: c1afd4573af58e6db3aa6d3bc46eec803a0fc2dd920055391720245c30c3579b
+  url: https://github.com/mzabrams/fars-cleaner/archive/refs/tags/v{{ version }}.tar.gz
+  sha256: 972f5d9f84fbfe87fcdf23b6c9a005b5851299f901e9d32748fb0eaf323f7276
 
 build:
   noarch: python
-  script: poetry install #{{ PYTHON }} -m pip install . -vv
+  script: {{ PYTHON }} -m pip install . -vv
   number: 0
 
-about:
-  home: https://pypi.org/project/fars-cleaner/
-  summary: A package for loading and preprocessing the NHTSA FARS crash database
-  license: BSD-3-Clause
-  license_file: LICENSE
-
 requirements:
+  build:
+    - python >=3.8
+    - poetry
   host:
     - pip
     - python >=3.8
     - poetry
-  build:
-    - python >=3.8
-    - poetry
   run:
-    - dask-core
+    - dask
+    - distributed
     - numpy
     - pandas >=1.4.0
     - pathlib
@@ -42,16 +34,28 @@ requirements:
     - requests
     - tqdm
     - thefuzz
+
 test:
   requires:
     - pip
     - hypothesis
     - pytest
     - poetry
+    - dask
+    - distributed
   imports:
     - fars_cleaner
+  source_files:
+    - tests/*.py
   commands:
-    - poetry run pytest
+    - pytest
+
+about:
+  home: https://pypi.org/project/fars-cleaner/
+  summary: A package for loading and preprocessing the NHTSA FARS crash database
+  license: BSD-3-Clause
+  license_family: BSD
+  license_file: LICENSE
 
 extra:
   recipe-maintainers:
