### 2022-10-18 7606ed1
Fix the test case in test_datasets.py'
