### 2022-10-18 d5ae73a
Fix python publish to only run on release
