+++ .github/workflows/pythonpublish.yml
@@ -5,7 +5,7 @@ name: Upload Python Package
 
 on:
   release:
-    types: [published]
+    types: [released]
 
 jobs:
   deploy:
