+++ meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "fars_cleaner" %}
-{% set version = "1.3.1" %}
+{% set version = "1.3.2" %}
 
 package:
   name: {{ name|lower }}
@@ -7,7 +7,7 @@ package:
 
 source:
   url: https://github.com/mzabrams/fars-cleaner/archive/refs/tags/v{{ version }}.tar.gz
-  sha256: c84d916de45ca2e8261618b6f52299e017f84a4710ab7b1ecc29eaf581cc8537
+  sha256: 693f81dfe294cd83e4e29aff5d841e5bc26f3e7579cd0e28514ffcf6cc86fd9a
 
 build:
   noarch: python
