+++ pyproject.toml
@@ -28,7 +28,7 @@ hypothesis = {version = "*", optional = true}
 
 [tool.poetry]
 name = "fars_cleaner"
-version = "1.3.2"
+version = "1.3.3"
 authors= [
     "Mitchell Abrams <mitchell.abrams@duke.edu>",
 ]
