+++ meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "fars_cleaner" %}
-{% set version = "1.3.2" %}
+{% set version = "1.3.3" %}
 
 package:
   name: {{ name|lower }}
@@ -7,7 +7,7 @@ package:
 
 source:
   url: https://github.com/mzabrams/fars-cleaner/archive/refs/tags/v{{ version }}.tar.gz
-  sha256: 693f81dfe294cd83e4e29aff5d841e5bc26f3e7579cd0e28514ffcf6cc86fd9a
+  sha256: 4fe007883baf484ad424e9b39ddbf3865a2637e0bfaca9e05b5c43b02f0c7daa
 
 build:
   noarch: python
