+++ .github/workflows/ci.yml
@@ -14,10 +14,10 @@ jobs:
       matrix:
         os: [ ubuntu-20.04, ubuntu-22.04, macos-12, macos-11, windows-latest]
         python-version: ['3.8', '3.9' ,'3.10']
-        exclude:
-          - os: windows-latest
-          - os: ubuntu-20.04
-          - os: ubuntu-22.04
+        #exclude:
+        #  - os: windows-latest
+        #  - os: ubuntu-20.04
+        #  - os: ubuntu-22.04
       fail-fast: false
     steps:
       - uses: actions/checkout@v3
