### 2022-10-18 mzabrams/fars-cleaner#3 @mzabrams
Pending PR acceptance on condo-forge, this will be resolved -- [conda-forge/staged-recipes#20792](https://github.com/conda-forge/staged-recipes/pull/20792). Currently, package is available on conda and pypi
