### 2022-10-18 #4678 @mzabrams
@majensen All but one of the relevant issues have been closed -- regarding issue [mzabrams/fars-cleaner#2](https://github.com/mzabrams/fars-cleaner/issues/3), the package is available for download via conda and pip, but is awaiting review by the conda-forge team to be accessed through that channel. Not sure what the timeline is for that team to move things on, but I think I've addressed everything else raised by @svburke and @ethanwhite
