### 2022-10-20 mzabrams/fars-cleaner#3 @mzabrams
Closing with [conda-forge/staged-recipes#20792](https://github.com/conda-forge/staged-recipes/pull/20792) -- fars_cleaner now available on PyPi, conda, and conda-forge
