### 2022-10-20 #4678 @majensen
Thanks @mzabrams - I'll do my editorial review of the paper itself (usu. pretty minor comments)- hope I can do it this weekend. Then the final steps to prep for the recommendation to the EIC are pretty rapid.
