### 2022-10-22 #4678 @majensen
@mzabrams - I like this software a lot, it is the kind of work that increases the usability of siloed data; I hope it goes viral and energizes your field.
2 very minor items: 1) Can you put the Statement of Need section just after the Summary; 2) can you add links to your proceedings in the refs. I see, for example, this link is available: http://www.ircobi.org/wordpress/downloads/irc22/pdf-files/2212.pdf.
