### 2022-10-22 #4678 @majensen
After do the above, can you create an open archive of your code repo, using Zenodo, FigShare or similiar? A short tutorial using Zenodo is [here](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). When this is done, please post the DOI you receive back in this thread. Thanks!
