+++ paper.bib
@@ -6,12 +6,14 @@
 	Organization = {{IRCOBI}},
 	Title = "{Female vs. Male Relative Fatality Risk in Fatal Crashes}",
 	booktitle = {Proceedings of 2020 International Research Council on the Biomechanics of Injury},
-	Year = {2020}}
+	Year = {2020},
+	url= {http://www.ircobi.org/wordpress/downloads/irc20/pdf-files/13.pdf}}
 
 
 @inproceedings{ircobi2022,
 	Author = {Abrams, Mitchell and Bass, C.R.},
 	Organization = {{IRCOBI}},
+	url = {http://www.ircobi.org/wordpress/downloads/irc22/pdf-files/2212.pdf},
 	Title = "{Female vs. Male Relative Risk of Body System Injuries in Fatal and Non-Fatal Crashes}",
 	booktitle = {Proceedings of 2022 International Research Council on the Biomechanics of Injury},
 	Year = {2022},
