### 2022-10-23 #4678 @mzabrams
@majensen Thanks! I think I'd already set it up with Zenodo -- DOI for the most recent release version (1.3.3) is 10.5281/zenodo.7221282. I've made those additions to the references (there isn't a publicly available repository of World Congress of Biomechanics Proceedings, unfortunately), and moved the statement of need as well. Let me know if there's anything else! I appreciate the kind words about this project!!
