### 2022-10-25 #4678 @majensen
Thanks @mzabrams - on the archive, JOSS prefers the archive name to be the name of the paper verbatim, and the archive authors to include all the paper authors. Would you be willing to change the title and author metadata on your current archive? Or alternatively, create a one-off archive that meets those desirables?
