### 2022-10-25 #4678 @mzabrams
@majensen I'll get the archive renamed with the underscore'd name this afternoon (may take a bit, as it will require a rename on PyPi as well, which is for the best as the conda and pip package names apparently differ -- I seem to have made a mistake there somewhere as pip is `fars-cleaner` and conda is `fars_cleaner`). Re: author metadata, currently Dr. Bass is an author on the paper (he is my PhD advisor), but not in the code metadata as he will not be a maintainer or responsible for correspondence/support/maintenance of the package in the future. Happy to make that change if it is the preference of the JOSS team, though
