### 2022-10-25 #4678 @majensen
Thanks @mzabrams - yeah, I've seen several pip packages where the dash is used in the name, but the package comes in with an underscore in code. Not sure that's a big problem, might be automatically translated even.
If you're ok putting your supervisor's name in the archive, then let's go ahead - might save a little editor-in-chief overhead.
