### 2022-10-25 #4678 @mzabrams
Sounds good -- I'll go ahead and do that! You're right it's not the biggest deal, just personally I think I'd like the pip/conda packages to have the same name to avoid future confusion with installation (although I'll look into whether that gets automatically translated).

Just need to add the name to `pyproject.toml`? Or any other locations?
