### 2022-10-25 #4678 @ethanwhite
> Dr. Bass is an author on the paper (he is my PhD advisor), but not in the code metadata as he will not be a maintainer or responsible for correspondence/support/maintenance of the package in the future. Happy to make that change if it is the preference of the JOSS team, though

@mzabrams - in cases like this we have included them as authors in the Zenodo archive (via a `.zenodo.json` file; which is what JOSS wants) but not listed them in setup.py, pypi, etc. I think this both allows the software paper and repo to match the author line while not implying direct code contributions (or ability to serve as a maintainer).
