### 2022-10-25 #4678 @mzabrams
> I'll look into whether that gets automatically translated

For the record, it appears pip automatically gets the same package whether I run `pip install fars_cleaner` or `pip install fars-cleaner`
