### 2022-10-25 d70c428
Remove Zenodo action
