### 2022-10-25 #4678 @mzabrams
Should be solved now -- the Zenodo DOI for the most recent version is [10.5281/zenodo.7249980](https://doi.org/10.5281/zenodo.7249980)

The DOI for all versions is [10.5281/zenodo.3735486](https://doi.org/10.5281/zenodo.3735486)
