+++ meta.yaml
@@ -1,13 +1,13 @@
 {% set name = "fars_cleaner" %}
-{% set version = "1.3.3" %}
+{% set version = "1.3.5" %}
 
 package:
   name: {{ name|lower }}
   version: {{ version }}
 
 source:
-  url: https://github.com/mzabrams/fars-cleaner/archive/refs/tags/v{{ version }}.tar.gz
-  sha256: 4fe007883baf484ad424e9b39ddbf3865a2637e0bfaca9e05b5c43b02f0c7daa
+  url: https://github.com/mzabrams/fars_cleaner/archive/refs/tags/v{{ version }}.tar.gz
+  sha256: eac566a3405f13fa8b329410a6956e79008d74921aa658e29263c32115826b69
 
 build:
   noarch: python
