### 2022-10-26 #4678 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
